Get detail vehicle & delete transaction dummy

+++ app/src/main/java/com/febrian/vehiclesales/data/entity/Vehicle.kt
@@ -4,7 +4,7 @@ open class Vehicle(
     open var id: Int = 0,
     open var name: String = "",
     open var year: Int = 0,
-    open var color: Long = 0,
+    open var color: Int = 0,
     open var price: Int = 0,
     open var isPurchased: Boolean = false
 )
